Remove QueryManager.queryIdsByName.
This map was essentially an index for quickly finding query IDs given a
query name. However, all the necessary information already exists in
this.queries (specifically, some of those queries have an observableQuery
property some of the time).

Although this index could have performance benefits in some cases,
refetching queries by name is relatively rare, and we can save a
considerable amount of code by not bothering to maintain the
queryIdsByName index.